[FEAT] 구글 로그인

+++ app/src/main/java/com/starters/yeogida/presentation/mypage/SettingFragment.kt
@@ -7,19 +7,19 @@ import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.webkit.WebViewClient
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
+import com.google.android.gms.auth.api.signin.GoogleSignIn
 import com.kakao.sdk.user.UserApiClient
 import com.starters.yeogida.R
 import com.starters.yeogida.YeogidaApplication
 import com.starters.yeogida.data.api.UserService
-import com.starters.yeogida.data.remote.common.TokenData
 import com.starters.yeogida.databinding.FragmentSettingBinding
 import com.starters.yeogida.network.YeogidaClient
 import com.starters.yeogida.presentation.common.CustomDialog
+import com.starters.yeogida.presentation.user.GoogleLoginObj
 import com.starters.yeogida.presentation.user.LoginActivity
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -76,9 +76,7 @@ class SettingFragment : Fragment() {
                     dismissDialog()
 
                     CoroutineScope(Dispatchers.IO).launch {
-                        val userAccessToken = dataStore.userAccessToken.first()
-                        val userRefreshToken = dataStore.userRefreshToken.first()
-                        withDrawUser(userAccessToken, userRefreshToken)
+                        withDrawUser()
                     }
                 }
                 setNegativeBtn("취소") {
@@ -88,7 +86,7 @@ class SettingFragment : Fragment() {
         }
     }
 
-    private suspend fun withDrawUser(userAccessToken: String, userRefreshToken: String) {
+    private suspend fun withDrawUser() {
         val withDrawResponse = userService.withDrawUser()
         Log.e("withDrawUser", "$withDrawResponse")
 
@@ -96,17 +94,39 @@ class SettingFragment : Fragment() {
             200 -> {
                 Log.e("withDrawUser", "code : 200, 회원 탈퇴 성공")
 
-                // DataStore 값 지우기
-                removeUserData()
+                when (dataStore.userLoginType.first()) {
+                    "kakao" -> {
+                        // 카카오 계정 연결 끊기
+                        UserApiClient.instance.unlink { error ->
+                            if (error != null) {
+                                Log.e("withDrawUser", "카카오 계정 연결 끊기 실패", error)
+                            } else {
+                                Log.i("withDrawUser", "카카오 계정 연결 끊기 성공")
+                            }
+                        }
+                    }
 
-                // 카카오 계정 연결 끊기
-                UserApiClient.instance.unlink { error ->
-                    if (error != null) {
-                        Log.e("withDrawUser", "카카오 계정 연결 끊기 실패", error)
-                    } else {
-                        Log.i("withDrawUser", "카카오 계정 연결 끊기 성공")
+                    "google" -> {
+                        // 구글 계정 탈퇴
+                        val mGoogleSignInClient =
+                            GoogleSignIn.getClient(requireActivity(), GoogleLoginObj.signInOption)
+                        mGoogleSignInClient.revokeAccess()
+                            .addOnCompleteListener {
+
+                            }
+                    }
+
+                    "naver" -> {
+
+                    }
+
+                    else -> {
+
                     }
                 }
+
+                // DataStore 값 지우기
+                removeUserData()
 
                 withContext(Dispatchers.Main) {
                     Toast.makeText(requireContext(), "탈퇴되었습니다.", Toast.LENGTH_SHORT).show()
@@ -115,12 +135,7 @@ class SettingFragment : Fragment() {
                 }
             }
 
-            403 -> { // 권한 없음, 만료된 토큰
-                Log.e("withDrawUser/Error", withDrawResponse.body()?.message.toString())
-                validateToken(userAccessToken, userRefreshToken)
-            }
-
-            404 -> {
+            403 -> {
                 // 존재하지 않는 회원
                 Log.e("UserWithDraw/Error", "${withDrawResponse.body()?.code}")
                 Log.e("UserWithDraw/Error", "${withDrawResponse.body()?.message}")
@@ -132,70 +147,15 @@ class SettingFragment : Fragment() {
         }
     }
 
-    private suspend fun validateToken(userAccessToken: String, userRefreshToken: String) {
-        val tokenResponse = userService.validateToken(
-            TokenData(
-                userAccessToken,
-                userRefreshToken
-            )
-        )
-
-        when (tokenResponse.code()) {
-            200 -> {
-                // 유효한 Access, Refresh
-                Log.e("MyPage/TokenValidation", "200")
-
-                // 토큰 값을 전달하여 탈퇴
-                withDrawUser(userAccessToken, userRefreshToken)
-            }
-
-            201 -> { // 만료된 Access
-                // newAccessToken 을 사용해서 회원 탈퇴 진행
-                Log.e("MyPage/TokenValidation", "201")
-                val newAccessToken = tokenResponse.body()?.data?.newAccessToken.toString()
-                val refreshToken = tokenResponse.body()?.data?.refreshToken.toString()
-
-                dataStore.saveUserToken(newAccessToken, refreshToken)
-                withDrawUser(newAccessToken, refreshToken)
-            }
-
-            403 -> { // Access, Refresh 둘 다 만료 or 비정상적인 형식의 토큰
-                Log.e("MyPage/TokenValidation", "403")
-                removeUserData()
-
-                withContext(Dispatchers.Main) {
-                    moveToLogin()
-                    Toast.makeText(
-                        requireContext(),
-                        "계정이 만료되었습니다. 다시 로그인 해주세요.",
-                        Toast.LENGTH_SHORT
-                    ).show()
-                }
-            }
-
-            else -> {
-                Log.e("MyPage/TokenValidation", "$tokenResponse")
-            }
-        }
-    }
-
     private fun setLogout() {
         viewModel.logOutEvent.observe(viewLifecycleOwner) {
-            UserApiClient.instance.logout { error ->
-                if (error != null) {
-                    Log.e("UserLogout", "로그아웃 실패", error)
-                } else {
-                    Log.i("UserLogout", "로그아웃 성공")
-                    Toast.makeText(requireContext(), "로그아웃 되었습니다.", Toast.LENGTH_SHORT).show()
-                    CoroutineScope(Dispatchers.IO).launch {
-                        // DataStore 값 지우기
-                        removeUserData()
-                    }
-
-                    // 로그인 화면으로
-                    moveToLogin()
-                }
+            CoroutineScope(Dispatchers.IO).launch {
+                // DataStore 값 지우기
+                removeUserData()
             }
+
+            // 로그인 화면으로
+            moveToLogin()
         }
     }
 
@@ -203,6 +163,7 @@ class SettingFragment : Fragment() {
         dataStore.saveIsLogin(false)
         dataStore.removeUserToken()
         dataStore.removeMemberId()
+        dataStore.removeLoginType()
     }
 
     private fun moveToLogin() {
